toke.c: Add guards that arent looking beyond end of input
Prior to this commit, we thought there was enough text in the input that
we could peek ahead of where we were by a few bytes.  This makes sure we
test that there is enough space before dereferencing.

I noticed these doing a code review, and have no examples of actual
failure.