yylex() Always check for UTF-8 well-formedness
(if it's UTF-8 that is)

While using gdb, I noticed that there were instances when input was
lexed without it being checked for being valid UTF-8.  Invalid UTF-8
can lead to core dumps, etc.  Previously we tried to avoid duplicate
checking by setting a flag, but this isn't working fully.  This commit
removes the flag and make the checking unconditional.  This may mean
that some of the input is checked twice.  What I learned when I took out
a whole pass of the regex compiler, is that the time for the compilation
step is essentially trivial compared to execution time, so the duplicate
checking won't be noticeable, and is far better than missing things
sometimes by not checking.