XXX tests PATCH: [perl #133998] regcomp assertion failure
This was caused in the blamed commit by me thinking I could simplify too
much.  So this commit puts back in checking specially for thing like
/[\s\S]/l, which has to match every possible code point no matter what
the locale.  Alternatively, I think I could have taken out the assertion
that was checking that this optimization happened at compile time, and
had it work at run time, but ISTR there being other issues with doing
that, and this, is better anyway, though likely rare in practice.